+++ FlacConvert/FolderConvert.h
@@ -9,7 +9,7 @@ class FolderConvert
 public:
 
     int ConverAllDirectories(const std::filesystem::path& directory, bool bAsync = false);
-    int GetFilesData(std::tuple<int, long, long>& scanInfo, const std::filesystem::path& directory, bool bAsync = false);
+    int GetFilesData(std::tuple<int, long, long long>& scanInfo, const std::filesystem::path& directory, bool bAsync = false);
 
 
     static bool IsFileConvertable(std::filesystem::path pathName);
